+++ x/subaccount/keeper/subaccount.go
@@ -1,7 +1,6 @@
 package keeper
 
 import (
-	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/sge-network/sge/x/subaccount/types"
@@ -51,90 +50,6 @@ func (k Keeper) GetSubAccountOwner(ctx sdk.Context, subaccountAddr sdk.AccAddres
 	store := ctx.KVStore(k.storeKey)
 	addr := store.Get(types.SubAccountKey(subaccountAddr))
 	return addr, addr != nil
-}
-
-// SetLockedBalances saves the locked balances of an account.
-func (k Keeper) SetLockedBalances(ctx sdk.Context, subAccountAddress sdk.AccAddress, lockedBalances []types.LockedBalance) {
-	store := ctx.KVStore(k.storeKey)
-
-	for _, lockedBalance := range lockedBalances {
-		amountBytes, err := lockedBalance.Amount.Marshal()
-		if err != nil {
-			panic(err)
-		}
-		store.Set(
-			types.LockedBalanceKey(subAccountAddress, lockedBalance.UnlockTime),
-			amountBytes,
-		)
-	}
-}
-
-// GetLockedBalances returns the locked balances of an account.
-func (k Keeper) GetLockedBalances(ctx sdk.Context, subAccountAddress sdk.AccAddress) []types.LockedBalance {
-	iterator := prefix.NewStore(ctx.KVStore(k.storeKey), types.LockedBalancePrefixKey(subAccountAddress)).Iterator(nil, nil)
-	defer iterator.Close()
-
-	var lockedBalances []types.LockedBalance
-	for ; iterator.Valid(); iterator.Next() {
-		unlockTime, err := sdk.ParseTimeBytes(iterator.Key())
-		if err != nil {
-			panic(err)
-		}
-
-		amount := new(sdk.Int)
-		err = amount.Unmarshal(iterator.Value())
-		if err != nil {
-			panic(err)
-		}
-		lockedBalances = append(lockedBalances, types.LockedBalance{
-			UnlockTime: unlockTime,
-			Amount:     *amount,
-		})
-	}
-
-	return lockedBalances
-}
-
-// GetUnlockedBalance returns the unlocked balance of an account.
-func (k Keeper) GetUnlockedBalance(ctx sdk.Context, subAccountAddress sdk.AccAddress) sdk.Int {
-	iterator := prefix.NewStore(ctx.KVStore(k.storeKey), types.LockedBalancePrefixKey(subAccountAddress)).
-		Iterator(nil, sdk.FormatTimeBytes(ctx.BlockTime()))
-
-	unlockedBalance := sdk.ZeroInt()
-	defer iterator.Close()
-
-	for ; iterator.Valid(); iterator.Next() {
-		amount := new(sdk.Int)
-		err := amount.Unmarshal(iterator.Value())
-		if err != nil {
-			panic(err)
-		}
-		unlockedBalance = unlockedBalance.Add(*amount)
-	}
-
-	return unlockedBalance
-}
-
-// SetBalance saves the balance of an account.
-func (k Keeper) SetBalance(ctx sdk.Context, subAccountAddress sdk.AccAddress, balance types.Balance) {
-	store := ctx.KVStore(k.storeKey)
-
-	bz := k.cdc.MustMarshal(&balance)
-	store.Set(types.BalanceKey(subAccountAddress), bz)
-}
-
-// GetBalance returns the balance of an account.
-func (k Keeper) GetBalance(ctx sdk.Context, subAccountAddress sdk.AccAddress) (types.Balance, bool) {
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.BalanceKey(subAccountAddress))
-	if bz == nil {
-		return types.Balance{}, false
-	}
-
-	balance := types.Balance{}
-	k.cdc.MustUnmarshal(bz, &balance)
-
-	return balance, true
 }
 
 func (k Keeper) IterateSubaccounts(ctx sdk.Context, cb func(subAccountAddress, subaccountOwner sdk.AccAddress) (stop bool)) {
